+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ### Added
     - `Boruta.Ecto.Admin.get_scopes_by_names/1`
     - `Boruta.Ecto.Client.grant_types/0`
+    - ability to insert/update clients with given id/secret
+    - insert/update a client inserts non existing authorized_scopes
 ## [2.0.0-rc.1]
   ### Fixed
     - better Ecto errors management